Refactored ProtectedMemberInFinalClass test cases
