+++ detekt-rules/src/test/kotlin/io/gitlab/arturbosch/detekt/rules/KT.kt
@@ -23,7 +23,8 @@ enum class Case(val file: String) {
 	Exceptions("/cases/Exceptions.kt"),
 	ExceptionRaisedInMethodsNegative("/cases/ExceptionRaisedInMethodsNegative.kt"),
 	ExceptionRaisedInMethodsPositive("/cases/ExceptionRaisedInMethodsPositive.kt"),
-	FinalClass("/cases/FinalClass.kt"),
+	FinalClassNegative("/cases/ProtectedMemberInFinalClassNegative.kt"),
+	FinalClassPositive("/cases/ProtectedMemberInFinalClassPositive.kt"),
 	FunctionReturningConstant("/cases/FunctionReturningConstant.kt"),
 	IteratorImplNegative("/cases/IteratorImplNegative.kt"),
 	IteratorImplPositive("/cases/IteratorImplPositive.kt"),
