+++ detekt-rules/src/test/resources/cases/ProtectedMemberInFinalClassNegative.kt
@@ -1,0 +1,31 @@
+@file:Suppress("unused", "RedundantVisibilityModifier", "ProtectedInFinal", "ConvertSecondaryConstructorToPrimary")
+
+package cases
+
+// should report 9 for protected = internal
+class NoProtectedMembersInFinalClass : BaseClass() {
+
+	private val i = 0
+
+	// should not report protected = private visibility
+	protected override val abstractProp = 0
+
+	// should not report protected = private visibility
+	protected override fun abstractFunction() {
+	}
+}
+
+abstract class BaseClass {
+
+	protected abstract val abstractProp: Int
+	protected abstract fun abstractFunction()
+
+	protected object InnerObject
+}
+
+sealed class SealedClass {
+
+	protected fun a() {}
+}
+
+
