+++ detekt-rules/src/test/resources/cases/ProtectedMemberInFinalClassPositive.kt
@@ -1,35 +1,22 @@
-@file:Suppress("unused")
+@file:Suppress("unused", "RedundantVisibilityModifier", "ProtectedInFinal", "ConvertSecondaryConstructorToPrimary")
 
 package cases
 
-// should report 9 for protected = internal
-class FinalClass : BaseClass {
+class ProtectedMemberInFinalClassPositive {
 
-	private val i = 0
 	protected var i1 = 0 // positive case
 
 	protected constructor(i1: Int) : super() { // positive case
 		this.i1 = i1
 	}
 
-	// should not report protected = private visibility
-	protected override val abstractProp = 0
-
-	// should not report protected = private visibility
-	protected override fun abstractFunction() {
-	}
-
-
-	protected fun function() {
-	} // positive case
+	protected fun function() {} // positive case
 
 	protected inner class InnerClass1 { // positive case
-
 		protected val i = 0 // positive case
 	}
 
 	inner class InnerClass2 {
-
 		protected val i = 0 // positive case
 	}
 
@@ -42,12 +29,7 @@ class FinalClass : BaseClass {
 	}
 }
 
-abstract class BaseClass {
-
-	protected abstract val abstractProp: Int
-	protected abstract fun abstractFunction()
-
-	protected object InnerObject
+abstract class ClassWithAbstractCompanionMembers {
 
 	protected companion object {
 		protected class A {
@@ -59,14 +41,8 @@ abstract class BaseClass {
 open class OpenClass {
 
 	inner class InnerClass {
-
 		protected val i = 0 // positive case
 	}
-}
-
-sealed class SealedClass {
-
-	protected fun a() {}
 }
 
 class FinalClassWithProtectedConstructor protected constructor() // positive case
